+++ viadot/tasks/azure_sql.py
@@ -8,7 +8,7 @@ from prefect import Task
 from prefect.tasks.secrets import PrefectSecret
 from prefect.utilities.tasks import defaults_from_attrs
 
-from ..exceptions import ValidationError
+
 from ..sources import AzureSQL
 from .azure_key_vault import AzureKeyVaultSecret
 
@@ -325,95 +325,3 @@ class AzureSQLToDF(Task):
 
         self.logger.info(f"Successfully downloaded data to a DataFrame.")
         return df
-
-
-class EnsureDFColumnOrder(Task):
-    """
-    Task for checking the order of columns in the loaded DF and in the SQL table into which the data from DF will be loaded.
-    If order is different then DF columns are reordered according to the columns of the SQL table.
-    """
-
-    def __init__(
-        self,
-        table: str = None,
-        schema: str = None,
-        if_exists: Literal["fail", "replace", "append", "delete"] = "replace",
-        df: pd.DataFrame = None,
-        credentials_secret: str = None,
-        vault_name: str = None,
-        *args,
-        **kwargs,
-    ):
-        self.credentials_secret = credentials_secret
-        self.vault_name = vault_name
-
-        super().__init__(name="run_ensure_df_column_order", *args, **kwargs)
-
-    def df_change_order(
-        self, df: pd.DataFrame = None, sql_column_list: List[str] = None
-    ):
-        df_column_list = list(df.columns)
-        if set(df_column_list) == set(sql_column_list):
-            df_changed = df.loc[:, sql_column_list]
-        else:
-            raise ValidationError(
-                "Detected discrepancies in number of columns or different column names between the CSV file and the SQL table!"
-            )
-
-        return df_changed
-
-    def sanitize_columns(self, df: pd.DataFrame = None):
-        """
-        Function to remove spaces at the end of column name.
-        Args:
-            df(pd.DataFrame): Dataframe to transform. Defaults to None.
-        """
-        for col in df.columns:
-            df = df.rename(columns={col: col.strip()})
-        return df
-
-    def run(
-        self,
-        table: str = None,
-        schema: str = None,
-        if_exists: Literal["fail", "replace", "append", "delete"] = "replace",
-        df: pd.DataFrame = None,
-        credentials_secret: str = None,
-        vault_name: str = None,
-    ):
-        """
-        Run a checking column order
-
-        Args:
-            table (str, optional): SQL table name without schema. Defaults to None.
-            schema (str, optional): SQL schema name. Defaults to None.
-            if_exists (Literal, optional): What to do if the table exists. Defaults to "replace".
-            df (pd.DataFrame, optional): Data Frame. Defaults to None.
-            credentials_secret (str, optional): The name of the Azure Key Vault secret containing a dictionary
-            with SQL db credentials (server, db_name, user, and password). Defaults to None.
-            vault_name (str, optional): The name of the vault from which to obtain the secret. Defaults to None.
-        """
-        credentials = get_credentials(credentials_secret, vault_name=vault_name)
-        azure_sql = AzureSQL(credentials=credentials)
-        df = self.sanitize_columns(df)
-        query = f"SELECT COLUMN_NAME FROM INFORMATION_SCHEMA.COLUMNS WHERE TABLE_SCHEMA = '{schema}' AND TABLE_NAME = '{table}'"
-        check_result = azure_sql.run(query=query)
-        if if_exists not in ["replace", "fail"]:
-            if if_exists == "append" and not check_result:
-                self.logger.warning("Aimed table doesn't exists.")
-                return
-            elif check_result is not []:
-                result = azure_sql.run(query=query)
-                sql_column_list = [table for row in result for table in row]
-                df_column_list = list(df.columns)
-
-                if sql_column_list != df_column_list:
-                    self.logger.warning(
-                        "Detected column order difference between the CSV file and the table. Reordering..."
-                    )
-                    df = self.df_change_order(df=df, sql_column_list=sql_column_list)
-                else:
-                    return df
-        else:
-            self.logger.info("The table will be replaced.")
-            return df
